phase-2-fullstack/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
psycopg2-binary
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1